Move MUC and stanza utils into shared and plugin-specific files
