+++ spec/chatbox.js
@@ -1007,19 +1007,17 @@ describe("Chatboxes", function () {
             const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit',
                   msg = mock.createChatMessage(_converse, sender_jid, 'This message will be unread');
 
-            const sent_stanzas = [];
-            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s));
-
             const view = await mock.openChatBoxFor(_converse, sender_jid)
-            spyOn(view.model, 'sendMarker').and.callThrough();
+            const sent_stanzas = [];
+            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s?.nodeTree ?? s));
             view.model.save('scrolled', true);
             await _converse.handleMessageStanza(msg);
             await u.waitUntil(() => view.model.messages.length);
             expect(view.model.get('num_unread')).toBe(1);
             const msgid = view.model.messages.last().get('id');
             expect(view.model.get('first_unread_id')).toBe(msgid);
-            await u.waitUntil(() => view.model.sendMarker.calls.count() === 1);
-            expect(sent_stanzas[0].nodeTree.querySelector('received')).toBeDefined();
+            await u.waitUntil(() => sent_stanzas.length);
+            expect(sent_stanzas[0].querySelector('received')).toBeDefined();
             done();
         }));
 
@@ -1031,15 +1029,14 @@ describe("Chatboxes", function () {
             await mock.waitForRoster(_converse, 'current', 1);
             const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
             const msg = mock.createChatMessage(_converse, sender_jid, 'This message will be read');
-            const sent_stanzas = [];
-            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s));
             await mock.openChatBoxFor(_converse, sender_jid);
+            const sent_stanzas = [];
+            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s?.nodeTree ?? s));
             const chatbox = _converse.chatboxes.get(sender_jid);
-            spyOn(chatbox, 'sendMarker').and.callThrough();
             await _converse.handleMessageStanza(msg);
             expect(chatbox.get('num_unread')).toBe(0);
-            await u.waitUntil(() => chatbox.sendMarker.calls.count() === 2);
-            expect(sent_stanzas[1].nodeTree.querySelector('displayed')).toBeDefined();
+            await u.waitUntil(() => sent_stanzas.filter(s => s.nodeName === 'message').length === 2);
+            expect(sent_stanzas[1].querySelector('displayed')).toBeDefined();
             done();
         }));
 
@@ -1053,12 +1050,10 @@ describe("Chatboxes", function () {
             const msgFactory = function () {
                 return mock.createChatMessage(_converse, sender_jid, 'This message will be unread');
             };
-
-            const sent_stanzas = [];
-            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s));
             await mock.openChatBoxFor(_converse, sender_jid);
+            const sent_stanzas = [];
+            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s?.nodeTree ?? s));
             const chatbox = _converse.chatboxes.get(sender_jid);
-            spyOn(chatbox, 'sendMarker').and.callThrough();
             _converse.windowState = 'hidden';
             const msg = msgFactory();
             _converse.handleMessageStanza(msg);
@@ -1066,8 +1061,8 @@ describe("Chatboxes", function () {
             expect(chatbox.get('num_unread')).toBe(1);
             const msgid = chatbox.messages.last().get('id');
             expect(chatbox.get('first_unread_id')).toBe(msgid);
-            await u.waitUntil(() => chatbox.sendMarker.calls.count() === 1);
-            expect(sent_stanzas[0].nodeTree.querySelector('received')).toBeDefined();
+            await u.waitUntil(() => sent_stanzas.filter(s => s.nodeName === 'message').length);
+            expect(sent_stanzas[0].querySelector('received')).toBeDefined();
             done();
         }));
 
@@ -1079,11 +1074,10 @@ describe("Chatboxes", function () {
             await mock.waitForRoster(_converse, 'current', 1);
             const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
             const msgFactory = () => mock.createChatMessage(_converse, sender_jid, 'This message will be unread');
-            const sent_stanzas = [];
-            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s));
             await mock.openChatBoxFor(_converse, sender_jid);
+            const sent_stanzas = [];
+            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s?.nodeTree ?? s));
             const chatbox = _converse.chatboxes.get(sender_jid);
-            spyOn(chatbox, 'sendMarker').and.callThrough();
             chatbox.save('scrolled', true);
             _converse.windowState = 'hidden';
             const msg = msgFactory();
@@ -1092,8 +1086,8 @@ describe("Chatboxes", function () {
             expect(chatbox.get('num_unread')).toBe(1);
             const msgid = chatbox.messages.last().get('id');
             expect(chatbox.get('first_unread_id')).toBe(msgid);
-            await u.waitUntil(() => chatbox.sendMarker.calls.count() === 1);
-            expect(sent_stanzas[0].nodeTree.querySelector('received')).toBeDefined();
+            await u.waitUntil(() => sent_stanzas.filter(s => s.nodeName === 'message').length === 1);
+            expect(sent_stanzas[0].querySelector('received')).toBeDefined();
             done();
         }));
 
@@ -1105,11 +1099,10 @@ describe("Chatboxes", function () {
             await mock.waitForRoster(_converse, 'current', 1);
             const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
             const msgFactory = () => mock.createChatMessage(_converse, sender_jid, 'This message will be unread');
-            const sent_stanzas = [];
-            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s));
             await mock.openChatBoxFor(_converse, sender_jid);
+            const sent_stanzas = [];
+            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s?.nodeTree ?? s));
             const chatbox = _converse.chatboxes.get(sender_jid);
-            spyOn(chatbox, 'sendMarker').and.callThrough();
             _converse.windowState = 'hidden';
             const msg = msgFactory();
             _converse.handleMessageStanza(msg);
@@ -1117,12 +1110,12 @@ describe("Chatboxes", function () {
             expect(chatbox.get('num_unread')).toBe(1);
             const msgid = chatbox.messages.last().get('id');
             expect(chatbox.get('first_unread_id')).toBe(msgid);
-            await u.waitUntil(() => chatbox.sendMarker.calls.count() === 1);
-            expect(sent_stanzas[0].nodeTree.querySelector('received')).toBeDefined();
+            await u.waitUntil(() => sent_stanzas.filter(s => s.nodeName === 'message').length === 1);
+            expect(sent_stanzas[0].querySelector('received')).toBeDefined();
             _converse.saveWindowState({'type': 'focus'});
             expect(chatbox.get('num_unread')).toBe(0);
-            await u.waitUntil(() => chatbox.sendMarker.calls.count() === 2);
-            expect(sent_stanzas[1].nodeTree.querySelector('displayed')).toBeDefined();
+            await u.waitUntil(() => sent_stanzas.filter(s => s.nodeName === 'message').length === 2);
+            expect(sent_stanzas[1].querySelector('displayed')).toBeDefined();
             done();
         }));
 
@@ -1134,11 +1127,10 @@ describe("Chatboxes", function () {
             await mock.waitForRoster(_converse, 'current', 1);
             const sender_jid = mock.cur_names[0].replace(/ /g,'.').toLowerCase() + '@montague.lit';
             const msgFactory = () => mock.createChatMessage(_converse, sender_jid, 'This message will be unread');
-            const sent_stanzas = [];
-            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s));
             await mock.openChatBoxFor(_converse, sender_jid);
+            const sent_stanzas = [];
+            spyOn(_converse.connection, 'send').and.callFake(s => sent_stanzas.push(s?.nodeTree ?? s));
             const chatbox = _converse.chatboxes.get(sender_jid);
-            spyOn(chatbox, 'sendMarker').and.callThrough();
             chatbox.save('scrolled', true);
             _converse.windowState = 'hidden';
             const msg = msgFactory();
@@ -1147,13 +1139,12 @@ describe("Chatboxes", function () {
             expect(chatbox.get('num_unread')).toBe(1);
             const msgid = chatbox.messages.last().get('id');
             expect(chatbox.get('first_unread_id')).toBe(msgid);
-            await u.waitUntil(() => chatbox.sendMarker.calls.count() === 1);
-            expect(sent_stanzas[0].nodeTree.querySelector('received')).toBeDefined();
+            await u.waitUntil(() => sent_stanzas.filter(s => s.nodeName === 'message').length === 1);
+            expect(sent_stanzas[0].querySelector('received')).toBeDefined();
             _converse.saveWindowState({'type': 'focus'});
             await u.waitUntil(() => chatbox.get('num_unread') === 1);
             expect(chatbox.get('first_unread_id')).toBe(msgid);
-            await u.waitUntil(() => chatbox.sendMarker.calls.count() === 1);
-            expect(sent_stanzas[0].nodeTree.querySelector('received')).toBeDefined();
+            expect(sent_stanzas[0].querySelector('received')).toBeDefined();
             done();
         }));
     });
