+++ spec/muc_messages.js
@@ -1,6 +1,6 @@
 /*global mock, converse */
 
-const { Promise, Strophe, $msg, $pres, sizzle, stanza_utils } = converse.env;
+const { Promise, Strophe, $msg, $pres, sizzle } = converse.env;
 const u = converse.env.utils;
 const original_timeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
 
@@ -621,85 +621,28 @@ describe("A Groupchat Message", function () {
         expect(view.el.querySelectorAll('.chat-msg').length).toBe(1);
 
         const msg_obj = view.model.messages.at(0);
-        const stanza = u.toStanza(`
+        let stanza = u.toStanza(`
+            <message xmlns="jabber:client"
+                     from="${msg_obj.get('from')}"
+                     to="${_converse.connection.jid}"
+                     type="groupchat">
+                <body>${msg_obj.get('message')}</body>
+                <stanza-id xmlns="urn:xmpp:sid:0"
+                           id="5f3dbc5e-e1d3-4077-a492-693f3769c7ad"
+                           by="lounge@montague.lit"/>
+                <origin-id xmlns="urn:xmpp:sid:0" id="${msg_obj.get('origin_id')}"/>
+            </message>`);
+        await view.model.handleMessageStanza(stanza);
+        await u.waitUntil(() => view.model.messages.last().get('received'));
+
+        stanza = u.toStanza(`
             <message xml:lang="en" to="romeo@montague.lit/orchard"
                      from="lounge@montague.lit/some1" type="groupchat" xmlns="jabber:client">
                 <received xmlns="urn:xmpp:receipts" id="${msg_obj.get('msgid')}"/>
                 <origin-id xmlns="urn:xmpp:sid:0" id="CE08D448-5ED8-4B6A-BB5B-07ED9DFE4FF0"/>
             </message>`);
-        spyOn(stanza_utils, "parseMUCMessage").and.callThrough();
         _converse.connection._dataRecv(mock.createRequest(stanza));
-        await u.waitUntil(() => stanza_utils.parseMUCMessage.calls.count() === 1);
         expect(view.el.querySelectorAll('.chat-msg').length).toBe(1);
-        expect(view.el.querySelectorAll('.chat-msg__receipt').length).toBe(0);
-        done();
-    }));
-
-    it("can cause a chat marker to be returned",
-        mock.initConverse(
-            ['rosterGroupsFetched'], {},
-            async function (done, _converse) {
-
-        await mock.waitForRoster(_converse, 'current');
-        const muc_jid = 'lounge@montague.lit';
-        await mock.openAndEnterChatRoom(_converse, muc_jid, 'romeo');
-        const view = _converse.api.chatviews.get(muc_jid);
-        const textarea = view.el.querySelector('textarea.chat-textarea');
-        textarea.value = 'But soft, what light through yonder airlock breaks?';
-        view.onKeyDown({
-            target: textarea,
-            preventDefault: function preventDefault () {},
-            keyCode: 13 // Enter
-        });
-        await new Promise(resolve => view.model.messages.once('rendered', resolve));
-        expect(view.el.querySelectorAll('.chat-msg').length).toBe(1);
-        expect(view.el.querySelector('.chat-msg .chat-msg__body').textContent.trim())
-            .toBe("But soft, what light through yonder airlock breaks?");
-
-        const msg_obj = view.model.messages.at(0);
-        let stanza = u.toStanza(`
-            <message xml:lang="en" to="romeo@montague.lit/orchard"
-                     from="lounge@montague.lit/some1" type="groupchat" xmlns="jabber:client">
-                <received xmlns="urn:xmpp:chat-markers:0" id="${msg_obj.get('msgid')}"/>
-            </message>`);
-        const stanza_utils = converse.env.stanza_utils;
-        spyOn(stanza_utils, "getChatMarker").and.callThrough();
-        _converse.connection._dataRecv(mock.createRequest(stanza));
-        await u.waitUntil(() => stanza_utils.getChatMarker.calls.count() === 1);
-        expect(view.el.querySelectorAll('.chat-msg').length).toBe(1);
-        expect(view.el.querySelectorAll('.chat-msg__receipt').length).toBe(0);
-
-        stanza = u.toStanza(`
-            <message xml:lang="en" to="romeo@montague.lit/orchard"
-                     from="lounge@montague.lit/some1" type="groupchat" xmlns="jabber:client">
-                <displayed xmlns="urn:xmpp:chat-markers:0" id="${msg_obj.get('msgid')}"/>
-            </message>`);
-        _converse.connection._dataRecv(mock.createRequest(stanza));
-        await u.waitUntil(() => stanza_utils.getChatMarker.calls.count() === 2);
-        expect(view.el.querySelectorAll('.chat-msg').length).toBe(1);
-        expect(view.el.querySelectorAll('.chat-msg__receipt').length).toBe(0);
-
-        stanza = u.toStanza(`
-            <message xml:lang="en" to="romeo@montague.lit/orchard"
-                     from="lounge@montague.lit/some1" type="groupchat" xmlns="jabber:client">
-                <acknowledged xmlns="urn:xmpp:chat-markers:0" id="${msg_obj.get('msgid')}"/>
-            </message>`);
-        _converse.connection._dataRecv(mock.createRequest(stanza));
-
-        await u.waitUntil(() => stanza_utils.getChatMarker.calls.count() === 3);
-        expect(view.el.querySelectorAll('.chat-msg').length).toBe(1);
-        expect(view.el.querySelectorAll('.chat-msg__receipt').length).toBe(0);
-
-        stanza = u.toStanza(`
-            <message xml:lang="en" to="romeo@montague.lit/orchard"
-                     from="lounge@montague.lit/some1" type="groupchat" xmlns="jabber:client">
-                <body>'tis I!</body>
-                <markable xmlns="urn:xmpp:chat-markers:0"/>
-            </message>`);
-        _converse.connection._dataRecv(mock.createRequest(stanza));
-        await u.waitUntil(() => stanza_utils.getChatMarker.calls.count() === 4);
-        await u.waitUntil(() => view.el.querySelectorAll('.chat-msg').length === 2);
-        expect(view.el.querySelectorAll('.chat-msg__receipt').length).toBe(0);
         done();
     }));
 });
